Add alien fire logic
If group is instantiated with fire ability, then a random location will be chosen for the alien to shoot at. the alien will charge and play an animation, then disappear when the bomb is fully charged. When the bomb reaches the bottom of the screen, it triggers a custom event which must be listened for