Update hitboxes and Add alternate player for iPad Playing
This is a bug I noticed as I was preparing to email Craig about the domain to access the game. The changes make hitboxes for bosses more correct on ipad screens and adds an alternate player for our cutscene videos because they arent playing video for some reason on ipads. The alternate player creates a new tab which plays the video from our cloud storage